add ES_load subcommand
